fixed some ts bugs index and connectWallet
